Single-sourcing the package version
The technique used is based on technique 1 (with the note for
setuptools >= 61.0.0) from

  https://packaging.python.org/en/latest/guides/single-sourcing-package-version/

but using 'src/diffannotator/config.py' file, rather than
'diffannotator/__init__.py' that the above mentioned resource suggests.
The idea for using .config / config.py comes from the "How do I share
global variables across modules?" section in Python's FAQ:
  https://docs.python.org/3/faq/programming.html#how-do-i-share-global-variables-across-modules

Possible alternatives would be to use the 'VERSION' file, like in
technique 4 (and using `version = {file = ["VERSION"]}` in 'pyproject.toml'),
or use technique 7: keep the version number in the tags of a version control
system (Git) instead of in the code, and automatically extract it from there
using setuptools_scm:

  https://setuptools-scm.readthedocs.io/en/stable/usage/

.
NOTE that config.py also includes the get_version() function
that tries to retrieve version number as _installed_.
